+++ api/pom.xml
@@ -59,12 +59,6 @@
             <artifactId>broadleaf-framework-web</artifactId>
         </dependency>
 
-        <!-- For Maven -->
-        <dependency>
-            <groupId>com.fasterxml.jackson.core</groupId>
-            <artifactId>jackson-databind</artifactId>
-            <version>2.13.4</version>
-        </dependency>
 
         <!-- this dependency is added for calling rest api of broadleaf such as for schemas or any wrapper implementation-->
         <dependency>
