+++ api/src/main/java/com/community/api/endpoint/customer/CustomerDTO.java
@@ -4,11 +4,13 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import org.springframework.stereotype.Component;
 
 @AllArgsConstructor
 @NoArgsConstructor
 @Setter
 @Getter
+@Component
 public class CustomerDTO {
     private String firstName,lastName,email,mobileNumber,username,role;
