+++ core/src/main/resources/runtime-properties/common-shared.properties
@@ -249,7 +249,7 @@ database.user=postgres
 
 # local database password
 
-database.password=postgres
+database.password=Alpha#01
 
 database.driver=org.postgresql.Driver
